+++ examples/resources/sapintegrationsuite_integration_flow_configuration/import.sh
@@ -1,0 +1,3 @@
+# "<flow_id>/<flow_version>". Import adopts every parameter the version has;
+# keys you leave out of the configuration afterwards are simply no longer managed.
+terraform import sapintegrationsuite_integration_flow_configuration.metering metering/1.0.3